+++ Ubuntu/core.sh
@@ -8,7 +8,7 @@ postInstallationLog "Gnome System Tools"
 logInfo "Available programs: network-admin, shares-admin, time-admin, users-admin"
 
 
-log "Installing packages to compress and extract different kind of files..."
+logInfo "Installing packages to compress and extract different kind of files..."
 sudo apt-get -fy install unace unrar zip unzip p7zip-full p7zip-rar sharutils rar uudeview mpack arj cabextract
 
 
@@ -16,16 +16,17 @@ preInstallationLog "GParted with NTFS support"
 sudo apt-get -fy install gparted ntfs-3g
 
 
-log "Adding repositories for Rhythmbox and its plugins..."
+logInfo "Adding repositories for Rhythmbox and its plugins..."
 sudo add-apt-repository ppa:fossfreedom/rhythmbox -y
 sudo add-apt-repository ppa:fossfreedom/rhythmbox-plugins -y
 sudo apt-get update
-log "Instaling latest Rhythmbox and its plugins..."
+logInfo "Instaling latest Rhythmbox and its plugins..."
 sudo apt-get -fy install rhythmbox rhythmbox-plugin-rhythmweb rhythmbox-plugin-equalizer rhythmbox-plugin-opencontainingfolder rhythmbox-plugin-llyrics
 
 
 preInstallationLog "Subdownloader, GMountISO, Freemind (a mind maps editor), Sound Converter, Steam client (will update on first run) and PlayOnLinux"
 sudo apt-get -fy install subdownloader gmountiso freemind soundconverter steam playonlinux
+postInstallationLog "Subdownloader, GMountISO, Freemind (a mind maps editor), Sound Converter, Steam client (will update on first run) and PlayOnLinux"
 
 
 if isUbuntu "$PLATFORM"; then
@@ -33,11 +34,11 @@ if isUbuntu "$PLATFORM"; then
     . ./installAll-15.04.sh
 fi
 
-log "Cleaning up..." &&
+logInfo "Cleaning up..." &&
 sudo apt-get -f install &&
 sudo apt-get autoremove &&
 sudo apt-get -y autoclean &&
 sudo apt-get -y clean
 
-log "Updating installed packages..."
+logInfo "Updating installed packages..."
 sudo apt-get upgrade