+++ Ubuntu/installAll-15.04.sh
@@ -2,16 +2,17 @@
 # Created by Nahuel Barrios.
 
 
-cd ~/Downloads/ || (echo "Failed cding into Downloads directory, exiting..." && exit)
+enterDirOrExit ~/Downloads
 downloadAndInstall "GetDeb" http://archive.getdeb.net/install_deb/getdeb-repository_0.1-1~getdeb1_all.deb
 downloadAndInstall "PlayDeb" http://archive.getdeb.net/install_deb/playdeb_0.3-1~getdeb1_all.deb
 
 
-log "Installing Dropbox, MongoDB and WebP command line tools..."
+preInstallationLog "Installing Dropbox, MongoDB and WebP command line tools..."
 sudo apt-get -fy install nautilus-dropbox mongodb curl webp
+postInstallationLog "Installing Dropbox, MongoDB and WebP command line tools..."
 
 
-cd ~/Downloads/ || (echo "Failed cding into Downloads directory, exiting..." && exit)
+enterDirOrExit ~/Downloads
 downloadAndInstall "Mega Sync client" https://mega.nz/linux/MEGAsync/xUbuntu_15.04/amd64/megasync-xUbuntu_15.04_amd64.deb
 downloadAndInstall "Mega Nautilus extension" https://mega.nz/linux/MEGAsync/xUbuntu_15.04/amd64/nautilus-megasync-xUbuntu_15.04_amd64.deb
 downloadAndInstall "TeamViewer" http://www.teamviewer.com/download/teamviewer_linux.deb