+++ createDirectoriesTree.sh
@@ -1,16 +1,16 @@
 #!/usr/bin/env bash
 # Created by Nahuel Barrios on 15/04/15. Contact me at: Barrios.Nahuel@gmail.com
 
-log "#### Creating directories tree ####"
+logInfo "#### Creating directories tree ####"
 
 cd ~/ &&
-mkdir .gradle ||
-mkdir Coding ||
+mkdir .gradle
+mkdir Coding
 cd Coding &&
-mkdir IDEs ||
-mkdir xDKs ||
-mkdir servers ||
-mkdir tools ||
-mkdir workspace-mine ||
-mkdir workspace-test ||
+mkdir IDEs
+mkdir xDKs
+mkdir servers
+mkdir tools
+mkdir workspace-mine
+mkdir workspace-test
 mkdir workspace-work