+++ modules/git.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Created by Nahuel Barrios on 17/3/16.
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
 if ! command -v git >/dev/null; then
 
@@ -14,13 +14,14 @@ if ! command -v git >/dev/null; then
         xcode-select --install
         postInstallationLog "Mac OS X Command line tools"
     fi
-
-    git config --global user.name "$USER_FULL_NAME" &&
-    git config --global user.email "$USER_EMAIL" &&
-    git config --global push.default simple &&
-    log "GIT configured successfully."
 else
-    logAlreadyInstalled "git"
+    logAlreadyInstalled "GIT"
 fi
+
+git config --global user.name "$USER_FULL_NAME" &&
+git config --global user.email "$USER_EMAIL" &&
+git config --global push.default matching &&
+git config --global core.autocrlf &&
+logSummary "GIT configured successfully."
 
 logProgramVersion "GIT" "$(git version)"