+++ modules/google-drive.sh
@@ -1,26 +1,24 @@
 #!/bin/bash
 # Created by Nahuel Barrios on 24/3/16.
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
-if ! command -v drive >/dev/null; then
+if isUbuntu "$PLATFORM" ; then
+    if ! command -v drive >/dev/null; then
 
-    preInstallationLog "Google Drive client"
-
-    if isUbuntu "$PLATFORM" ; then
-        log "Adding repositories for Google Drive client..."
+        preInstallationLog "Google Drive client"
+        logInfo "Adding repositories for Google Drive client..."
 
         sudo add-apt-repository ppa:twodopeshaggy/drive
         sudo apt-get update
         sudo apt-get -fy install drive
+
+        postInstallationLog "Google Drive client"
     else
-        go get -u github.com/odeke-em/drive/cmd/drive
+        logAlreadyInstalled "Google Drive Client"
     fi
 
-    postInstallationLog "Google Drive client"
+    logProgramVersion "Google Drive client" "$(drive version)"
 else
-    logAlreadyInstalled "Google Drive Client"
+  logSummary "Remember to download the Google Drive client from: https://www.google.com/drive/download/"
 fi
-
-logInfo "See https://github.com/odeke-em/drive#usage to mount your Google Drive directory on your local filesystem"
-logProgramVersion "Google Drive client" "$(drive version)"