+++ modules/homebrew.sh
@@ -2,7 +2,7 @@
 # Created by Nahuel Barrios on 18/3/16.
 # shellcheck disable=SC1091
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
 preInstallationLog "Homebrew"
 if ! command -v brew >/dev/null; then
@@ -13,15 +13,14 @@ else
   logAlreadyInstalled "Homebrew"
 fi
 
-log "Updating Homebrew formulas..."
+logInfo "Updating Homebrew formulas..."
 brew update
 
-log "Verifying the Homebrew installation..."
+logInfo "Verifying the Homebrew installation..."
 if brew doctor; then
   postInstallationLog "Homebrew"
 else
-  log "Your Homebrew installation reported some errors or warnings."
-  logError "Review the Homebrew messages to see if any action is needed."
+  logError "Your Homebrew installation reported some errors or warnings, review the Homebrew messages to see if any action is needed."
 fi
 
 logProgramVersion "Homebrew" "$(brew --version)"