+++ modules/jq.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 # Created by Nahuel Barrios on 25/3/16.
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
-preInstallationLog "jq json beautifer"
+preInstallationLog "jq JSON beautifer"
 
 if isUbuntu "$PLATFORM" ; then
     sudo apt-get -fy install jq
@@ -11,4 +11,4 @@ else
     brew install jq
 fi
 
-postInstallationLog "jq json beautifer"
+postInstallationLog "jq JSON beautifer"