+++ modules/mixxx.sh
@@ -1,18 +1,18 @@
 #!/bin/bash
 # Created by Nahuel Barrios on 25/3/16.
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
 if isUbuntu "$PLATFORM" ; then
 
     preInstallationLog "Mixxx DJ Software"
 
-    log "Adding repositories for Mixxx DJ software..."
+    logInfo "Adding repositories for Mixxx DJ software..."
     sudo add-apt-repository ppa:mixxx/mixxx -y
     sudo apt-get update
     sudo apt-get -fy install mixxx libportaudio2
 
     postInstallationLog "Mixxx DJ Software"
 else
-    logInfo "Download Mixxx from: http://www.mixxx.org/download/"
+    logSummary "Download Mixxx from: http://www.mixxx.org/download/"
 fi