+++ modules/sublime-text.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 # Created by Nahuel Barrios on 25/3/16.
 
-cd "${CURRENT_DIR}" || (echo "Failed cding into EFU's execution directory, exiting..." && exit)
+enterDirOrExit "${CURRENT_DIR}"
 
 if isUbuntu "$PLATFORM" ; then
-    log "Adding repositories for Sublime Text 3..."
+    preInstallationLog "Sublime Text 3"
+
+    logInfo "Adding repositories for Sublime Text 3..."
     sudo add-apt-repository ppa:webupd8team/sublime-text-3 -y
     sudo apt-get update
 
-    log "Instaling Sublime Text 3"
+    logInfo "Instaling Sublime Text 3"
     sudo apt-get -fy install sublime-text-installer
+
+    postInstallationLog "Sublime Text 3"
 else
-    logInfo "Download Sublime Text from: https://www.sublimetext.com/3"
+    logSummary "Download Sublime Text from: https://www.sublimetext.com/3"
 fi